noise model for *selected* aperture, not optimal
Optimizing the apertures is an impossible task because we do not know
the field angles, so we cannot possibly properly model the PSFs and
figure out how much flux from neighboring stars can be easily resolved.

The commit gives `src/noise_model.py`, the default noise model, a
slightly-modified baseline set of assumptions.

* Rather than trying to compute an optimal aperture size, we assume an
  analytic function that specifies the aperture size as a function of
  TESS mag.

* Sky noise is only zodiacal (no "unresolved background star fluid")

* Shot noise is computed based on TESS-band zero point (not Teff and
  Cousins I)

* Systematic and readout noise are computed the same way as before.

All neighbor stars will be accounted for in the TSWG's priority scheme
through the method discussed by Stassun+ 2017 in the
https://arxiv.org/abs/1706.00495 (v3, posted Wed 28 June 2017).

This scheme, to be executed by Ryan Oelkers, will compensate for the
stellar companions that are currently not included in this noise model.

-----

The previous set of assumptions (which actually does compute an
"optimal" aperture size, but making different assumptions about
contamination) have been archived in
`src/noise_model_optimal_aperture.py`